Add --expose-stack flag.
The observation can contain the first two elements of the stack (aka the values
of the list elements pointed by the pointers stored in the stack). One can also
train everything without exposing the stack.

Two booleans were used to specify the ordering between pointers. This was done
to avoid to expose the pointers values directly.