Fix bug inside the quicksort observation generation.
If we do not want to expose the values of the pointers and we do not want
to expose the stack content either then we just need to expose the boolean
values.

The issue was caused by np.concatenate() being called with a zero-dimensional
array. This means that np.concatenate() was called with only on np.array as
argument. Obviously, this is pointless since we cannot concatenate just one
array. Hence the error.