Update quicksort environment to expand observations.
The observations now contain also information about the ordering properties
of the elements inside the list. This however can be seen as cheating, since
we give explicit information to the model about the notion of "ordering".

However, it can be noted that the network has still no idea of what "greater"
and "smaller" means. Therefore it will still need to learn what that kind of
information means.